refactor: Preserve alias members path by re-aliasing members instead of returning target's members
